number operations #1

+++ lib/src/number/real.dart
@@ -4,6 +4,7 @@ part of number;
 ///
 abstract class Real extends Number {
   const Real.constant() : super.constant();
+
   Real();
 
   /// All Real subclasses must be able to provide their value as a [dart:core] [num].
@@ -12,29 +13,44 @@ abstract class Real extends Number {
   int toInt();
   double toDouble();
 
+  @override
   bool get isInfinite => value == double.INFINITY || value == double.NEGATIVE_INFINITY;
+
+  @override
   bool get isNaN => value == double.NAN;
+
+  @override
   bool get isNegative => value < 0;
 
+  /// Addition operator.
+  ///
+  /// [addend] is expected to be a `num` or `Number`.  If it is not it
+  /// will be assumed to be 0.
+  ///
+  @override
   Number operator +(addend) {
     if (addend is num) return new Double(value + addend);
     if (addend is Real) return new Double(value + addend.value);
     if (addend is Complex) return new Complex(new Double(addend.real.toDouble() + value), addend.imag);
     if (addend is Imaginary) return new Complex(this, addend);
+    if (addend is Precise) return addend + this;
     return this;
   }
 
+  @override
   Number operator -() => new Double(-value);
 
+  @override
   Number operator -(subtrahend) {
     if (subtrahend is num) return new Double(value - subtrahend);
     if (subtrahend is Real) return new Double(value - subtrahend.value);
     if (subtrahend is Complex) return new Complex(new Double(value - subtrahend.real.value), -(subtrahend.imag));
     if (subtrahend is Imaginary) return new Complex(this, -subtrahend);
-
+    if (subtrahend is Precise) return (-subtrahend) + this;
     return this;
   }
 
+  @override
   Number operator *(multiplier) {
     if (multiplier is num) {
       num product = multiplier * value;
@@ -44,11 +60,13 @@ abstract class Real extends Number {
     if (multiplier is Real) return this * multiplier.value;
     if (multiplier is Complex) return new Complex(multiplier.real * value, multiplier.imaginary * value);
     if (multiplier is Imaginary) return new Imaginary(multiplier.value * value);
+    if (multiplier is Precise) return multiplier * this;
 
     // Treat multiplier as 0
     return Integer.zero;
   }
 
+  @override
   Number operator /(divisor) {
     if (divisor is num) return new Double(value / divisor);
     if (divisor is Real) return new Double(value / divisor.value);
@@ -58,7 +76,9 @@ abstract class Real extends Number {
       Number aOverc2d2 = this / c2d2;
 
       return new Complex(aOverc2d2 * divisor.real, new Imaginary(aOverc2d2 * divisor.imaginary.value * -1.0));
-    } else if (divisor is Imaginary) return new Imaginary((this / divisor.value) * -1);
+    }
+    if (divisor is Imaginary) return new Imaginary((this / divisor.value) * -1);
+    if (divisor is Precise) return (new Precise.num(this.value)) / divisor;
 
     // Treat divisor as 0
     return Double.infinity;
@@ -69,6 +89,7 @@ abstract class Real extends Number {
   /// When dividing by an [Imaginary] or [Complex] number, the result will contain an imaginary component.
   /// The imaginary component is *not* truncated; only the real portion of the result is truncated.
   ///
+  @override
   Number operator ~/(divisor) {
     if (divisor == 0) return Double.infinity;
     if (divisor is num) return new Integer(value ~/ divisor);
@@ -80,9 +101,8 @@ abstract class Real extends Number {
       return new Complex(
           (aOverc2d2 * divisor.real).truncate(), new Imaginary(aOverc2d2 * divisor.imaginary.value * -1.0));
     }
-    if (divisor is Imaginary) {
-      return new Imaginary(((this / divisor.value) * -1).truncate());
-    }
+    if (divisor is Imaginary) return new Imaginary(((this / divisor.value) * -1).truncate());
+    if (divisor is Precise) return (new Precise.num(this.value) / divisor).truncate();
 
     // Treat divisor as 0
     return Double.infinity;
@@ -92,6 +112,7 @@ abstract class Real extends Number {
   ///
   /// When dividing by an [Imaginary] or [Complex] number...
   ///
+  @override
   Number operator %(divisor) {
     if (divisor is num) return new Double(value % divisor);
     if (divisor is Real) return new Double(value % divisor.value);
@@ -99,6 +120,8 @@ abstract class Real extends Number {
       // (a + 0i) / (c + di) = (ac - adi) / (c^2 + d^2)
       Number c2d2 = (divisor.real ^ 2.0) + (divisor.imaginary.value ^ 2.0);
       Number aOverc2d2 = this / c2d2;
+
+      //TODO Real % operator incomplete
 
       //return new Complex(aOverc2d2 * divisor.real, new Imaginary(aOverc2d2 * divisor.imaginary.value * -1.0));
       //TODO
@@ -121,15 +144,14 @@ abstract class Real extends Number {
   /// for bitwise XOR operations on [int]s.  The [Integer] class provides the [bitwiseXor] method
   /// as a substitute.
   ///
+  @override
   Number operator ^(exponent) {
     if (exponent is num) {
       num raised = Math.pow(value, exponent);
       if (raised is int) return new Integer(raised);
       return new Double(raised);
     }
-    if (exponent is Real) {
-      return new Double(Math.pow(value, exponent.value));
-    }
+    if (exponent is Real) return new Double(Math.pow(value, exponent.value));
     if (exponent is Complex) {
       // a^(b+ic) = a^b * ( cos(c * ln(a)) + i * sin(c * ln(a)) )
       Number coeff = this ^ exponent.real;
@@ -141,35 +163,60 @@ abstract class Real extends Number {
       double clna = (exponent.value * Math.log(value)).toDouble();
       return new Complex(Math.cos(clna) as Real, new Imaginary(Math.sin(clna)));
     }
+    if (exponent is Precise) return (new Precise.num(value)) ^ exponent;
     return Double.one;
   }
 
+  @override
   bool operator >(obj) {
     if (obj is num) return value > obj;
     if (obj is Real) return value > obj.value;
     if (obj is Imaginary) return value > 0;
     if (obj is Complex) return this > obj.real;
+    if (obj is Precise) return new Precise.num(this.value) > obj;
     return this > 0;
   }
 
+  @override
   bool operator >=(obj) => (this == obj) ? true : this > obj;
 
+  @override
   bool operator <(obj) => !(this >= obj);
 
+  @override
   bool operator <=(obj) => !(this > obj);
 
+  @override
   Number abs() => value >= 0 ? this : value is int ? new Integer(value.abs()) : new Double(value.abs().toDouble());
 
-  Integer ceil() => new Integer(value.ceil());
+  @override
+  Number ceil() => new Integer(value.ceil());
 
-  Integer truncate() => new Integer(value.truncate());
+  @override
+  Number floor() => new Integer(value.floor());
 
+  @override
+  Number round() => new Integer(value?.round());
+
+  @override
+  Number truncate() => new Integer(value.truncate());
+
+  @override
   Number reciprocal() {
     if (value == 0) return Double.NaN;
     else if (value == 1) return Integer.one;
     else return new Double(1.0 / value);
   }
 
+  @override
+  Number remainder(divisor) {
+    num div = divisor is num ? divisor : divisor is Number ? divisor.toDouble() : 0;
+    var rem = value.remainder(div);
+    if (rem is int) return new Integer(rem);
+    return new Double(rem);
+  }
+
+  @override
   String toString() {
     return "${value}";
   }
